Last changes of the week
The layout is a mess still (sorry) but am trying to make sure the diff pairs are nicely positioned on the stack up. There are a handful of discrete components unconnected which I will need to go through the schematic carefully to place properly, but the general layout of components is getting there.